Move ContractTableController (it did not land where it was supposed to...)
